--- src/frontend/src/lib/components/homepage/ScrollTo.ts
@@ -1,7 +1,0 @@
-export const scrollTo = (target: string): void => {
-  const el = document.querySelector(target);
-  if (!el) return;
-  el.scrollIntoView({
-    behavior: "smooth",
-  });
-};
