+++ src/frontend/src/lib/utils/scroll.ts
@@ -1,0 +1,7 @@
+export const scroll = (target: string): void => {
+  const el: Element | null = document.querySelector(target);
+  if (!el) return;
+  el.scrollIntoView({
+    behavior: "smooth",
+  });
+};
